requirements.txt
fastapi
uvicorn
streamlit
requests
pydantic
python-dotenv
google-generativeai
crewai
crewai_tools
